% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betan_cfps.R
\name{betan}
\alias{betan}
\title{Betan: Computation method of SES proposed by Betancourt, L, 2016}
\usage{
betan(
  faminc,
  familysize,
  itn,
  edu_m,
  method = "raw",
  data = "CFPS",
  povertyline = 12060,
  increase = 4180
)
}
\arguments{
\item{faminc}{Family income}

\item{familysize}{An integer indicating the number of people in the household}

\item{itn}{(Optional) income-to-need ratio, only supplied when ITN is pre-calculated}

\item{edu_m}{Mother's education}

\item{method}{\code{'raw'}(default) indicate calculate SES by family income,
education and family size. \code{'itn'} indicate calculate SES by pre-calculated ITN}

\item{data}{A character indicate ‘CFPS' dataset or 'PSID'}

\item{povertyline}{Only used in 'PSID' dataset, the value, by default is 12060}

\item{increase}{Only used in 'PSID' dataset, the value, by default is 4180}
}
\value{
A single column of SES
}
\description{
Betan: Computation method of SES proposed by Betancourt, L, 2016
}
