% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFPS_child.R
\docType{data}
\name{CFPS_child}
\alias{CFPS_child}
\title{CFPS2010 child dataset for SES reproduction}
\format{
## `CFPS_child`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.isss.pku.edu.cn/cfps/>
}
\usage{
CFPS_child
}
\description{
A subset of data from the CFPS survey
}
\keyword{datasets}
