% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcde.R
\name{mcde}
\alias{mcde}
\title{McDermott:Computation method of SES proposed by McDermott, 2019 (CFPS only)}
\usage{
mcde(occup_m, edu_m, occup_f, edu_f, data = "CFPS")
}
\arguments{
\item{occup_m}{mother's occupation}

\item{edu_m}{mother's education}

\item{occup_f}{father's occupation}

\item{edu_f}{father's education}

\item{data}{A character indicate 'CFPS' dataset or 'PSID'.And in this case, only CFPS is applicable}
}
\value{
A single column of SES
}
\description{
McDermott:Computation method of SES proposed by McDermott, 2019 (CFPS only)
}
