% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moog.R
\name{moog}
\alias{moog}
\title{Moog: Computation method of SES proposed by Moog, 2008}
\usage{
moog(faminc, edu_m, data = "CFPS")
}
\arguments{
\item{faminc}{Family income}

\item{edu_m}{Mother's education}

\item{data}{A character indicate 'CFPS' dataset or 'PSID'}
}
\value{
A single column of SES
}
\description{
Moog: Computation method of SES proposed by Moog, 2008
}
