% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSID_adult.R
\docType{data}
\name{PSID_adult}
\alias{PSID_adult}
\title{PSID adult dataset for SES reproduction}
\format{
## `PSID_adult`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://psidonline.isr.umich.edu/>
}
\usage{
PSID_adult
}
\description{
A subset of data from the CFPS survey
}
\keyword{datasets}
