% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiu.R
\name{qiu}
\alias{qiu}
\title{Qiu: Computation method of SES proposed by Qiu, 2017 (CFPS & PSID)}
\usage{
qiu(faminc, data = "CFPS")
}
\arguments{
\item{faminc}{household income}

\item{data}{A character indicate 'CFPS'  or 'PSID' dataset}
}
\value{
A single column of SES
}
\description{
Qiu: Computation method of SES proposed by Qiu, 2017 (CFPS & PSID)
}
