% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romeo_a.R
\name{romeo_a}
\alias{romeo_a}
\title{Computation method of SES proposed by Romeo, 2018a (CFPS only)}
\usage{
romeo_a(occup_m, edu_m, data = "CFPS")
}
\arguments{
\item{occup_m}{mother's occupation}

\item{edu_m}{mother's education}

\item{data}{A character indicate 'CFPS' dataset or 'PSID'.And in this case, only CFPS is applicable,because occupation data is not available from PSID}
}
\value{
A single column of SES
}
\description{
Computation method of SES proposed by Romeo, 2018a (CFPS only)
}
