% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romeo_b.R
\name{romeo_b}
\alias{romeo_b}
\title{Romeo_b: Computation method of SES proposed by Romeo, 2018b (CFPS & PSID)}
\usage{
romeo_b(faminc, edu_m, edu_f, data = "CFPS")
}
\arguments{
\item{faminc}{Family income}

\item{edu_m}{Mother's education (If only one side of the parent's data is available, it is possible to fill in the list with all NA)}

\item{edu_f}{Father's education (If only one side of the parent's data is available, it is possible to fill in the list with all NA)}

\item{data}{A character indicate 'CFPS'  or 'PSID' dataset}
}
\value{
A single column of SES
}
\description{
Romeo_b: Computation method of SES proposed by Romeo, 2018b (CFPS & PSID)
}
